projects/fraud-detection/requirements.txt
# Core Data Science - Versiones compatibles con Python 3.13
pandas==2.3.2
numpy==2.3.3
scikit-learn==1.7.2
imbalanced-learn==0.14.0

# SQL y Procesamiento
sqlalchemy==2.0.43
pyarrow==21.0.0

# Visualización
matplotlib==3.10.6
seaborn==0.13.2
plotly==6.3.0

# Machine Learning - Versiones más recientes
xgboost==3.0.5
lightgbm==4.6.0
shap==0.48.0

# Dashboard
streamlit==1.50.0

# API REST
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.11.7
python-multipart==0.0.20

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# Monitoring & Alerts
requests==2.32.5
scipy==1.16.2

# Desarrollo
jupyter==1.1.1
ipykernel==6.30.1
notebook==7.4.5
black==25.9.0
pytest==8.4.2
pytest-cov==4.1.0
httpx==0.28.1

# Utilidades
python-dotenv==1.1.1
pyyaml==6.0.2
tqdm==4.67.1
joblib==1.5.2

# macOS específicas
appnope==0.1.4
